Remove d8-contrib-modules/cloudflarephpsdk library
d8-contrib-modules/cloudflarephpsdk is a library necessary
for drupal/cloudflare.
It would be installed as part of the standard - optional build.
Removing from the codebase because it should only be
installed if the client requires it, so that we are not
trying to maintain module code we are not utilizing.